test_work/requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.1.14
django-filter==2.4.0 
djangorestframework==3.12.4 
flake8==4.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
psycopg2-binary 
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.9
gunicorn==20.0.4 
drf-yasg==1.20.0
